refactor | fireball skills are now fireballs
